Alter to Long

+++ app/src/main/java/com/example/csi5175/register.kt
@@ -34,7 +34,7 @@ class register : AppCompatActivity() {
         submitButton.setOnClickListener{
             val addres = Address(country.text.toString(),state.text.toString(),zipcode.text.toString(),city.text.toString(),street.text.toString())
 
-            var phonenumber = phone.text.toString().toInt()
+            var phonenumber = phone.text.toString().toLong()
 
             db?.userDao()?.insert(User(email = email.text.toString(), password = password.text.toString(),firstName = firstName.text.toString(),lastName = lastName.text.toString(),phone = phonenumber, address = addres, history = null, favorite = null, cart = null))
             val user = db?.userDao()?.getUserByEmailAndPassword(email.text.toString(),password.text.toString())
